# Improve dedupe version performance